Change the implementation of IndlingResources
To avoid duplicated code, this pull request changes the implementation of the OkHttpIdlingResources, abstracting all the code to the Okir abstract class.
