+++ library/src/main/java/net/rafaeltoledo/okir/OkHttp3IdlingResource.java
@@ -1,18 +1,12 @@
 package net.rafaeltoledo.okir;
 
-import android.support.test.espresso.IdlingResource;
-
 import java.io.IOException;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import okhttp3.Interceptor;
 import okhttp3.Response;
 
-public class OkHttp3IdlingResource implements Interceptor, IdlingResource {
-
+public class OkHttp3IdlingResource extends Okir implements Interceptor {
     private final String[] strictUrls;
-    private AtomicInteger busy = new AtomicInteger(0);
-    private ResourceCallback callback;
 
     public OkHttp3IdlingResource() {
         this.strictUrls = new String[0];
@@ -42,20 +36,7 @@ public class OkHttp3IdlingResource implements Interceptor, IdlingResource {
     }
 
     @Override
-    public String getName() {
-        return "net.rafaeltoledo.okhttpidlingresource.OkHttp3IdlingResource";
-    }
-
-    @Override
-    public boolean isIdleNow() {
-        if (busy.get() == 0 && callback != null) {
-            callback.onTransitionToIdle();
-        }
-        return busy.get() == 0;
-    }
-
-    @Override
-    public void registerIdleTransitionCallback(ResourceCallback callback) {
-        this.callback = callback;
+    protected String getIdlingResourceName() {
+        return OkHttp3IdlingResource.class.getCanonicalName();
     }
 }
