+++ library/src/main/java/net/rafaeltoledo/okir/OkHttpIdlingResource.java
@@ -1,18 +1,13 @@
 package net.rafaeltoledo.okir;
 
-import android.support.test.espresso.IdlingResource;
-
 import com.squareup.okhttp.Interceptor;
 import com.squareup.okhttp.Response;
 
 import java.io.IOException;
-import java.util.concurrent.atomic.AtomicInteger;
 
-public class OkHttpIdlingResource implements Interceptor, IdlingResource {
+public class OkHttpIdlingResource extends Okir implements Interceptor {
 
     private final String[] strictUrls;
-    private AtomicInteger busy = new AtomicInteger(0);
-    private ResourceCallback callback;
 
     public OkHttpIdlingResource() {
         this.strictUrls = new String[0];
@@ -42,20 +37,7 @@ public class OkHttpIdlingResource implements Interceptor, IdlingResource {
     }
 
     @Override
-    public String getName() {
-        return "net.rafaeltoledo.okhttpidlingresource.OkHttpIdlingResource";
-    }
-
-    @Override
-    public boolean isIdleNow() {
-        if (busy.get() == 0 && callback != null) {
-            callback.onTransitionToIdle();
-        }
-        return busy.get() == 0;
-    }
-
-    @Override
-    public void registerIdleTransitionCallback(ResourceCallback callback) {
-        this.callback = callback;
+    protected String getIdlingResourceName() {
+        return OkHttpIdlingResource.class.getCanonicalName();
     }
 }
